power: bq27541: Check twice before reporting 0% SOC
In very rare cases, the fuel-gauge can suddenly and erroneously report a 0%
charge, causing the device to immediately shut down if the SOC is below the
critical point (41%, as the fuel-gauge is allowed to drop the SOC as fast as
it wants below 41%).

Double check before reporting 0% SOC in order to prevent the device from
randomly shutting down.

Signed-off-by: Sultanxda <sultanxda@gmail.com>